FEAT: IMPROVEMENTS USER API
I have added more validations to the registry. In addition, I have created two more endpoints: user edition, in which they can only change their first name and last name, and user deletion, both must be validated by the token assigned to that user that is offered when registering and when logging in.